refactor: #28 change button folder conventions
